+++ react/components/ProductName/index.js
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import ContentLoader from 'react-content-loader'
+import classNames from 'classnames'
 
 /**
  * Name component. Show name and relevant SKU information of the Product Summary
@@ -21,21 +22,33 @@ class ProductName extends Component {
     brandName: PropTypes.string,
     /** Show brand name */
     showBrandName: PropTypes.bool,
-    /** Display large font */
-    large: PropTypes.bool,
     /** Component and content loader styles */
     styles: PropTypes.object,
+    /** Classes to apply to elements of the component */
+    classes: PropTypes.shape({
+      root: PropTypes.string,
+      brandName: PropTypes.string,
+      skuName: PropTypes.string,
+      productReference: PropTypes.string,
+      rootLoader: PropTypes.string
+    })
   }
 
   static defaultProps = {
-    large: false,
     showBrandName: false,
     showProductReference: false,
     showSku: false,
+    classes: {
+      root: null,
+      brandName: null,
+      skuName: null,
+      productReference: null,
+      rootLoader: null
+    }
   }
 
   static Loader = (loaderProps = {}) => (
-    <div className="vtex-product-name vtex-product-name-loader pt5 overflow-hidden">
+    <div className={classNames('vtex-product-name vtex-product-name-loader', loaderProps.className)}>
       <ContentLoader
         style={{
           width: '100%',
@@ -65,43 +78,33 @@ class ProductName extends Component {
   render() {
     const {
       name,
+      classes,
       skuName,
-      brandName,
-      large,
-      showBrandName,
-      showProductReference,
       showSku,
+      brandName,
+      showBrandName,
       productReference,
+      showProductReference,
     } = this.props
-
-    let brandClasses = 'vtex-product-name__brand f5'
-    let skuClasses = 'vtex-product-name__sku f6'
-
-    if (large) {
-      brandClasses += ' vtex-product-name__brand--large f2'
-      skuClasses += ' vtex-product-name__sku--large'
-    }
 
     if (!name) {
       return (
-        <ProductName.Loader
-          {...this.props.styles}
-          brandClasses={brandClasses}
-          skuClasses={skuClasses}
-        />
+        <ProductName.Loader className={classes.rootLoader} {...this.props.styles} />
       )
     }
 
     return (
-      <div className="vtex-product-name overflow-hidden c-on-base">
-        <div className={brandClasses}>
+      <div className={classNames('vtex-product-name', classes.root)}>
+        <span className={classNames('vtex-product-name__brand', classes.brandName)}>
           {name} {showBrandName && brandName && `- ${brandName}`}
-        </div>
-        {showSku && <div className={skuClasses}>{skuName}</div>}
+        </span>
+        {showSku && skuName && (
+          <span className={classNames('vtex-product-name__sku', classes.skuName)}>{skuName}</span>
+        )}
         {showProductReference && productReference && (
-          <div className="vtex-product-name__product-reference pt3 f7 ttu">
+          <span className={classNames('vtex-product-name__product-reference', classes.productReference)}>
             {`REF: ${productReference}`}
-          </div>
+          </span>
         )}
       </div>
     )