+++ CHANGELOG.md
@@ -8,9 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Add several props to style the `ProductPrice` component.
+- Add `classes` prop to ProductName and remove `large` prop.
 
 ### Removed
-- `Footer` component.
+- `Header` from `vtex.store-components`.
 
 ## [2.6.4] - 2018-11-20
 ### Fixed